pyproject.toml
[tool.poetry]
name = "recipies"
version = "0.1.0"
description = ""
authors = ["Daniel Vo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
